requirements.txt
qdrant-client==1.11.3
sentence-transformers==3.0.1
torch>=2.1.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
uvicorn==0.30.6
fastapi==0.115.0
pydantic==2.8.2
